+++ src/components/Layout/Header/Header.js
@@ -361,6 +361,14 @@ class Header extends Component {
                                         <DropdownItem tag={RouterNavLink} exact to={ROUTES.TEMPLATES}>
                                             Templates
                                         </DropdownItem>
+                                        <DropdownItem
+                                            tag={RouterNavLink}
+                                            exact
+                                            to={ROUTES.CONTRIBUTION_EDITOR}
+                                            onClick={e => this.requireAuthentication(e, ROUTES.CONTRIBUTION_EDITOR)}
+                                        >
+                                            Contribution editor
+                                        </DropdownItem>
                                     </DropdownMenu>
                                 </UncontrolledButtonDropdown>
 
@@ -399,16 +407,30 @@ class Header extends Component {
 
                             <SearchForm placeholder="Search..." />
 
-                            <RequireAuthentication
-                                component={Button}
-                                color={!this.state.isHomePageStyle ? 'primary' : 'light'}
-                                className="mr-3 pl-4 pr-4 flex-shrink-0"
-                                tag={Link}
-                                to={ROUTES.ADD_PAPER.GENERAL_DATA}
-                            >
-                                <FontAwesomeIcon className="mr-1" icon={faPlus} />
-                                Add paper
-                            </RequireAuthentication>
+                            <UncontrolledButtonDropdown className="mr-3 flex-shrink-0">
+                                <RequireAuthentication
+                                    component={Button}
+                                    color={!this.state.isHomePageStyle ? 'primary' : 'light'}
+                                    className="pl-4 pr-4"
+                                    tag={Link}
+                                    to={ROUTES.ADD_PAPER.GENERAL_DATA}
+                                >
+                                    <FontAwesomeIcon className="mr-1" icon={faPlus} />
+                                    Add paper
+                                </RequireAuthentication>
+
+                                <DropdownToggle
+                                    split
+                                    color={!this.state.isHomePageStyle ? 'primary' : 'light'}
+                                    className="px-2"
+                                    style={{ marginLeft: 1 }}
+                                />
+                                <DropdownMenu right>
+                                    <DropdownItem tag={RouterNavLink} exact to={ROUTES.ADD_COMPARISON}>
+                                        Add comparison
+                                    </DropdownItem>
+                                </DropdownMenu>
+                            </UncontrolledButtonDropdown>
 
                             {!!this.props.user && (
                                 <div>